#5004 Move template in the deep clean.

+++ fabric8-maven-plugin/src/main/java/io/fabric8/maven/CleanMojo.java
@@ -59,7 +59,9 @@ public class CleanMojo extends ApplyMojo {
             openShiftClient.imageStreams().delete();
             openShiftClient.buildConfigs().delete();
             openShiftClient.deploymentConfigs().delete();
-            openShiftClient.templates().delete();
+            if (deep) {
+                openShiftClient.templates().delete();
+            }
         }
     }
 }
